### 2019-09-21 #1754 @whedon
[PRE REVIEW]: tabula: An R Package for Analysis, Seriation and Visualization of Archaeological Count Data
**Submitting author:** @nfrerebeau (<a href="http://orcid.org/0000-0001-5759-4944">Nicolas Frerebeau</a>)
**Repository:** <a href="https://github.com/nfrerebeau/tabula" target ="_blank">https://github.com/nfrerebeau/tabula</a>
**Version:** v1.3.0
