### 2019-09-21 #1754 @kthyng
@oliviaguest generally you don't need too much specific expertise, and you're on the lower end of editor load, but I know I've asked you to take on submissions this week!  I can try another editor first though.

@karthik this submission looks to be more in your area of expertise. Would you edit this submission?
