### 2019-09-23 #1754 @oliviaguest
Yes, @kthyng. I am starting a new job and moving house this week and next week. Is it OK to have a bit of a delay?
