### 2019-09-23 #1754 @danielskatz
Yes, let's go ahead and do that - I'll assign it you but with the recognition that there will be a ~2-week delay.

Dan (newly on-duty as JOSS AEiC this week)
