### 2019-09-25 742dbf1
Increase tolerance on noLD platforms when testing refined CA (#7)
