+++ NAMESPACE
@@ -48,6 +48,8 @@ exportMethods(similarity)
 exportMethods(test_diversity)
 exportMethods(test_fit)
 exportMethods(turnover)
+importFrom(ca,ca)
+importFrom(ggplot2,ggplot)
 importFrom(methods,new)
 importFrom(methods,setClass)
 importFrom(methods,setClassUnion)
