### 2019-10-01 4e58982
Fix a typo in the text and a misspelled name in the references (openjournals/joss-reviews#1754)
