+++ paper.bib
@@ -91,13 +91,13 @@
   date = {2002},
   author = {O'Brien, Michael J and Lyman, R. Lee}
 }
-@article{pietrie1899,
+@article{petrie1899,
   title = {Sequences in Prehistoric Remains},
   volume = {29},
   journaltitle = {Journal of the Anthropological Institute},
   date = {1899},
   pages = {295-301},
-  author = {Pietrie, W. M. F.}
+  author = {Petrie, W. M. F.}
 }
 @inproceedings{poblome2003,
   langid = {english},
