+++ vignettes/bibliography.bib
@@ -459,13 +459,13 @@
 	pages = {285-307},
 	author = {Peet, R. K.}
 }
-@article{pietrie1899,
+@article{petrie1899,
   title = {Sequences in Prehistoric Remains},
   volume = {29},
   journaltitle = {Journal of the Anthropological Institute},
   date = {1899},
   pages = {295-301},
-  author = {Pietrie, W. M. F.}
+  author = {Petrie, W. M. F.}
 }
 @report{philippe2017,
   langid = {english},
