+++ man/test.Rd
@@ -79,3 +79,4 @@ Feder, A. F., Kryazhimskiy, S. & Plotkin, J. B. (2014). Identifying
 \author{
 N. Frerebeau
 }
+\concept{test}
