### 2019-10-02 681ae5a
Change version to 1.3.0.9000
