+++ NEWS.Rmd
@@ -16,9 +16,9 @@ knitr::opts_chunk$set(
 )
 ```
 
-<!-- ## tabula `r utils::packageVersion("tabula")` (`r Sys.Date()`) -->
+## tabula `r utils::packageVersion("tabula")` (`r Sys.Date()`)
 
-## tabula 1.3.0
+## tabula 1.3.0 (release date: 2019-09-20)
 ### New classes and methods
 * `Matrix` this S4 class is now the superclass of all matrix-like classes.
 * `AbundanceMatrix` this virtual S4 class is defined as the superclass of `CountMatrix`, `FrequencyMatrix` and `IncidenceMatrix`.
