### 2019-10-19 #1821 @whedon
[REVIEW]: tabula: An R Package for Analysis, Seriation and Visualization of Archaeological Count Data
**Submitting author:** @nfrerebeau (<a href="http://orcid.org/0000-0001-5759-4944">Nicolas Frerebeau</a>)
**Repository:** <a href="https://github.com/nfrerebeau/tabula" target ="_blank">https://github.com/nfrerebeau/tabula</a>
**Version:** v1.4.0
