### 2019-10-19 #1821 @oliviaguest
Hey @izaromanowska and @soodoku!

This is where the reviewing proper happens. You need to go through the checklist and feel free to post questions here and open review issues in the submission repository: https://github.com/nfrerebeau/tabula (If you do please link to them here to help keep track.)
