### 2019-10-31 #1821 @izaromanowska
Hi @oliviaguest, I've done the review following the checklist. All works well as far as I can tell and is impressively well described. Do you need a written review as well?
One tiny suggestion is to mention in the readme that khroma needs to be installed and loaded. Probably unnecessary for most users but it's the only dependency that kicked off when checking examples on a freshly installed RStudio.
