### 2019-10-31 4ba29aa
Add plot methods for assemblage diversity size comparison (#9)
