+++ R/refine_diversity.R
@@ -22,10 +22,9 @@ setMethod(
 setMethod(
   f = "refine_evenness",
   signature = signature(object = "CountMatrix"),
-  definition = function(object, method = "shannon",
-                        prob = NULL, level = 0.80, n = 1000) {
+  definition = function(object, prob = NULL, level = 0.80, n = 1000) {
 
-    fun <- switch_evenness(method)
+    fun <- function(x) { evennessShannon(x, zero.rm = FALSE) }
     replicates <- refine_diversity(x = object, method = fun,
                                    prob = prob, level = level, n = n)
     return(replicates)
@@ -41,27 +40,31 @@ refine_diversity <- function(x, method, prob = NULL, level = 0.80, n = 1000) {
   size <- max(rowSums(x))
   sample_sizes <- seq_len(size * 1.05)
 
-  replicates <- vapply(
-    X = sample_sizes,
-    FUN = function(x, n, prob, level, method) {
-      types <- as.factor(seq_along(prob))
-      replicates <- stats::rmultinom(n, size = x, prob = prob)
-      R <- apply(X = replicates, MARGIN = 2, FUN = method)
 
-      # Confidence interval as described in Kintigh 1989
-      # k <- (1 - level) / 2
-      # conf <- quantile(R, probs = c(k, k + level))
+  simulate_diversity <- function(x, n, prob, level, method) {
+    replicates <- stats::rmultinom(n, size = x, prob = prob)
+    R <- apply(X = replicates, MARGIN = 2, FUN = method)
 
-      # Confidence interval
-      conf <- mean(R) + stats::sd(R) * stats::qnorm(level / 2) * c(-1, 1)
+    # Confidence interval as described in Kintigh 1989
+    # k <- (1 - level) / 2
+    # conf <- quantile(R, probs = c(k, k + level))
 
-      c(mean(R), conf)
-    },
-    FUN.VALUE = numeric(3),
-    n, prob, level, method
-  )
-  replicates <- cbind(sample_sizes, t(replicates))
-  colnames(replicates) <- c("size", "mean", "lower", "upper")
+    # Confidence interval
+    conf <- mean(R) + stats::sd(R) * stats::qnorm(level / 2) * c(-1, 1)
 
-  return(replicates)
+    c(mean(R), conf)
+  }
+  loop_args <- list(X = sample_sizes, FUN = simulate_diversity,
+                    n = n, prob = prob, level = level, method = method)
+  loop_fun <- if (requireNamespace("pbapply", quietly = TRUE)) {
+    pbapply::pblapply
+  } else {
+    lapply
+  }
+  simulated <- do.call(loop_fun, loop_args)
+  simulated <- do.call(rbind, simulated)
+  simulated <- cbind(sample_sizes, simulated)
+  colnames(simulated) <- c("size", "mean", "lower", "upper")
+
+  return(simulated)
 }
