### 2019-10-31 efd82b0
Add methods and rename refine_diversity to simulate_diversity (#9)
