+++ DESCRIPTION
@@ -99,10 +99,10 @@ Collate:
     'plot-spot.R'
     'plot-time.R'
     'refine.R'
-    'refine_diversity.R'
     'seriation.R'
     'seriate.R'
     'show.R'
+    'simulate.R'
     'statistics.R'
     'tabula-package.R'
     'test.R'
