+++ NAMESPACE
@@ -41,14 +41,14 @@ exportMethods(plot_time)
 exportMethods(rarefaction)
 exportMethods(refine)
 exportMethods(refine_dates)
-exportMethods(refine_evenness)
-exportMethods(refine_richness)
 exportMethods(refine_seriation)
 exportMethods(richness)
 exportMethods(seriate)
 exportMethods(seriate_correspondence)
 exportMethods(seriate_reciprocal)
 exportMethods(similarity)
+exportMethods(simulate_evenness)
+exportMethods(simulate_richness)
 exportMethods(test_diversity)
 exportMethods(test_fit)
 exportMethods(turnover)

[…]
