### 2019-10-31 2d760e6
Install and load suggested packages in the README (openjournals/joss-reviews#1821)
