+++ README.Rmd
@@ -1,8 +1,5 @@
 ---
 output: github_document
-header-includes:
-   - \usepackage{amsmath}
-   - \usepackage{amssymb}
 editor_options: 
   chunk_output_type: console
 ---
@@ -40,7 +37,7 @@ An easy way to examine archaeological count data. This package provides a conven
 
 ## Installation
 
-You can install the released version of `tabula` from [CRAN](https://CRAN.R-project.org) with:
+You can install the released version of **tabula** from [CRAN](https://CRAN.R-project.org) with:
 
 ```{r cran-installation, eval = FALSE}
 install.packages("tabula")
@@ -55,13 +52,23 @@ remotes::install_github("nfrerebeau/tabula")
 
 ## Usage
 
-```{r packages}
+```{r packages-installation, echo=FALSE}
+# Install missing packages (if any)
+packages <- c("khroma", "magrittr")
+if (!all(packages %in% utils::installed.packages()))
+  utils::install.packages(packages)
+```
+
+```{r packages-load}
 # Load packages
 library(tabula)
+
+library(khroma)
+library(ggplot2)
 library(magrittr)
 ```
 
-`tabula` provides a set of S4 classes that extend the basic `matrix` data type. These new classes represent different special types of matrix.
+**tabula** provides a set of S4 classes that extend the basic `matrix` data type. These new classes represent different special types of matrix.
 
 * Abundance matrix:
     * `CountMatrix` represents count data,
@@ -87,7 +94,7 @@ quali <- IncidenceMatrix(data = sample(0:1, 100, TRUE),
                          nrow = 10, ncol = 10)
 ```
 
-`tabula` uses coercing mechanisms (with validation methods) for data type conversions:
+**tabula** uses coercing mechanisms (with validation methods) for data type conversions:
 
 ```{r coerce}
 ## Create a count matrix
@@ -110,9 +117,9 @@ C <- as_incidence(A1)
 D <- as_occurrence(A1)
 ```
 
-Several types of graphs are available in `tabula` which uses `ggplot2` for plotting informations. This makes it easy to customize diagrams (e.g. using themes and scales).
+Several types of graphs are available in **tabula** which uses [**ggplot2**](https://github.com/tidyverse/ggplot2) for plotting informations. This makes it easy to customize diagrams (e.g. using themes and scales).
 
-Spot matrix[^1] allows direct examination of data:
+Spot matrix^[Adapted from Dan Gopstein's original [idea](https://dgopstein.github.io/articles/spot-matrix/).] allows direct examination of data:
 
 ```{r plot-occ, fig.width=6, fig.height=4, fig.align="center"}
 # Plot co-occurrence of types
@@ -231,7 +238,7 @@ mississippi %>%
   khroma::scale_colour_iridescent()
 ```
 
-The Frequency Increment Test can be used to assess the detection and quantification of selective processes in the archaeological record[^2].
+The Frequency Increment Test can be used to assess the detection and quantification of selective processes in the archaeological record^[Adapted from Ben Marwick's original [idea](https://github.com/benmarwick/signatselect/).].
 
 ```{r plot-time, fig.width=7, fig.height=3.5, fig.align="center"}
 ## Keep only decoration types that have a maximum frequency of at least 50
@@ -248,7 +255,4 @@ plot_time(merzbach_count, highlight = "FIT", roll = TRUE) +
 ```
 
 ## Contributing
-Please note that the `tabula` project is released with a [Contributor Code of Conduct](https://github.com/nfrerebeau/tabula/blob/master/.github/CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
-
[…]
