### 2019-11-11 #1821 @nfrerebeau
As suggested by @izaromanowska, I have modified the README so that khroma is installed and loaded.
