### 2019-11-15 74606b1
grammar + style suggestions
the paper is still hard to read and still has too much jargon. i would love to be of help to craft something that is clearer.
