+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'tabula: An R Package for Analysis, Seriation and Visualization of Archaeological Count Data'
+title: 'tabula: An R Package for Analysis, Seriation, and Visualization of Archaeological Count Data'
 tags:
   - archaeology
   - matrix seriation
@@ -19,22 +19,24 @@ bibliography: paper.bib
 ---
 
 # Background
-Detection and quantification of material and cultural variations in time and space are two key methodological issues in archaeology. Whether for intrasite or regional investigations, these issues require the construction of reliable chronologies and the quantitative description of archaeological assemblages (i. e. archaeological sites or intrasite units, each described as a set of $p$ types of objects).
 
-Building chronologies involve distinguishing between relative (that provide only a chronological sequence) and absolute dating methods (that yield a calendric indication) [@obrien2002]. Among the relative dating methods, matrix seriation is a long-established method in archaeology since its first formulation by @petrie1899 and has allowed the construction of reference chronologies [@ihm2005]. For a set $X$ of $n$ archaeological assemblages, the seriation problem comes down to discovering on $X$ an order inferred as chronological. This approach relies on a set of well-defined statistical and archaeological assumptions [@dunnell1970] and the use of *a priori* information (e. g. absolute dates or stratigraphical constraints: @poblome2003) allows the analysis of chronological patterns in a socio-economic or cultural perspective (e. g. @bellanger2012, @lipo2015).
+Detecting and quantifying material and cultural variations in time and space are important methodological issues in archaeology. To solve these issues, we need to construct reliable chronologies and quantitative descriptions of archaeological assemblages, i. e. archaeological sites or intrasite units, each described as a set of $p$ different objects.
 
-The quantitative analysis of archaeological assemblages can thus be carried out in a synchronic (e. g. diversity measurements) or diachronic (e. g. evolutionary studies: selection process, patterns of cultural transmission, etc.) way. These approaches cover a wide range of applications and have led to the development of a multitude of statistical models, but have not been systematically implemented to enable the deployment of reproducible workflows.
+Building chronologies involves distinguishing between relative (providing only a chronological sequence) and absolute dating methods (that yield calendric indicators) [@obrien2002]. Within relative dating, matrix seriation is a long-established method---it was first formulated by @petrie1899---and has allowed the construction of reference chronologies [@ihm2005]. For a set $X$ of $n$ archaeological assemblages, the seriation problem comes down to discovering in $X$ an order inferred as chronological. This approach relies on a set of well-defined statistical and archaeological assumptions [@dunnell1970]. It uses *a priori* information, e.g., absolute dates or stratigraphical constraints: @poblome2003, and allows the analysis of chronological patterns in a socio-economic or cultural perspective, e.g., @bellanger2012, @lipo2015.
+
+The quantitative analysis of archaeological assemblages can thus be carried out in a synchronic, e.g., diversity measurements, or diachronic, e.g., evolutionary studies: selection process, patterns of cultural transmission, etc., way. These approaches cover a wide range of applications and have led to the development of a multitude of statistical models, but none have been systematically implemented to enable the deployment of reproducible workflows.
 
 # Summary
-`tabula` attempts to provide a convenient and reproducible toolkit for analysis, seriation and visualization of archaeological count data (artifacts, faunal remains, etc.).
 
[…]
