### 2019-11-16 #1821 @soodoku
I have created a few issues. None are blockers but would love to get @nfrerebeau's eyes on them.

https://github.com/nfrerebeau/tabula/issues

once that's done, we can move fast.
