### 2019-11-19 #1821 @nfrerebeau
@soodoku thank you for all these constructive remarks. I will look in detail before the weekend and make the necessary changes.
