### 2019-11-19 103626d
Remove irrelevant note in the CountMatrix class documentation (#11)
