### 2019-11-19 959c71b
Add replacement methods for *Matrix classes that test the validity of the resulting object (fixes #13)
