+++ NAMESPACE
@@ -3,7 +3,9 @@
 export(CountMatrix)
 export(IncidenceMatrix)
 exportMethods("[")
+exportMethods("[<-")
 exportMethods("[[")
+exportMethods("[[<-")
 exportMethods("set_coordinates<-")
 exportMethods("set_dates<-")
 exportMethods("set_epsg<-")
