### 2019-11-19 6fb44c1
Add an alias to access the FrequencyMatrix class documentation (fixes #14)
