+++ NAMESPACE
@@ -61,9 +61,12 @@ importFrom(ggplot2,geom_point)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,labs)
 importFrom(ggplot2,scale_x_log10)
+importFrom(methods,as)
+importFrom(methods,callNextMethod)
 importFrom(methods,new)
 importFrom(methods,setClass)
 importFrom(methods,setClassUnion)
 importFrom(methods,setGeneric)
 importFrom(methods,setMethod)
+importFrom(methods,validObject)
 importFrom(rlang,.data)
