### 2019-11-19 af55811
Make error message more explicit (#11)
