+++ R/plot-prepare.R
@@ -171,6 +171,7 @@ prepare_spot <- function(object, threshold = NULL, diag = TRUE) {
   }
   if (is_square(object)) {
     max_value <- unique(diag(object))
+    if (max_value == 0) max_value <- max(data$value)
     data <- cbind.data.frame(max = max_value, data)
   }
   if (is.function(threshold)) {
