+++ _pkgdown.yml
@@ -1,4 +1,3 @@
-destination: docs
 url: http://tabula.archaeo.science
 development:
   mode: auto
