### 2019-11-20 1c6df27
Add an S4 class to store diversity index values and jackknife/bootstrap estimates (#15)
