+++ R/AllGenerics.R
@@ -355,24 +355,23 @@ setGeneric(
 #' Heterogeneity and Evenness
 #'
 #' @description
-#'  \code{diversity} returns an heterogeneity or dominance index.
+#'  \code{index_heterogeneity} returns an heterogeneity or dominance index.
 #'
-#'  \code{evenness} returns an evenness measure.
+#'  \code{index_evenness} returns an evenness measure.
 #' @param object A \eqn{m \times p}{m x p} matrix of count data (typically
 #'  a \linkS4class{CountMatrix} object).
-#' @param method A \code{\link{character}} string or vector of strings
-#'  specifying the index to be computed (see details).
-#'  Any unambiguous substring can be given.
-#' @param simplify A \code{\link{logical}} scalar: should the result be
-#'  simplified to a matrix? The default value, \code{FALSE}, returns a list.
-#' @param size A \code{\link{numeric}} vector giving the sample size.
-#' @param prob A length-\eqn{p} \code{\link{numeric}} vector giving the of
-#'  probability of the \eqn{p} taxa/types (see below). If \code{NULL} (the
-#'  default), probabilities are estimated from the whole dataset.
+#' @param method A \code{\link{character}} string specifying the index to be
+#'  computed (see details). Any unambiguous substring can be given.
+#' @param jackknife A \code{\link{logical}} scalar: should jackknifed estimates
+#'  be computed?
+#' @param bootstrap A \code{\link{logical}} scalar: should boostraped estimates
+#'  be computed?
+#' @param simulate A \code{\link{logical}} scalar: simulated assemblages
+#'  be computed?
 #' @param level A length-one \code{\link{numeric}} vector giving the
 #'  confidence level.
 #' @param n A non-negative \code{\link{integer}} giving the number of bootstrap
-#' replications (see below).
+#' replications.
 #' @param ... Further arguments to be passed to internal methods.
 #' @details
 #'  \emph{Diversity} measurement assumes that all individuals in a specific
@@ -428,8 +427,17 @@ setGeneric(
 #'  so that an increase in the value of the index accompanies a decrease in
 #'  diversity.
 #' @return
-#'  If \code{simplify} is \code{FALSE}, then \code{diversity} and
-#'  \code{evenness} return a list (default), else return a matrix.
+#'  \code{index_heterogeneity} and \code{index_evenness} return a
+#'  \linkS4class{DiversityIndex}.
+#'
+#'  \code{simulate_evenness} returns a \code{\link{data.frame}} with the
+#'  following columns:
+#'  \describe{
+#'   \item{size}{The sample sizes.}
+#'   \item{mean}{The boostraped mean.}
+#'   \item{lower}{The lower boundary of the boostraped confidence interval.}
+#'   \item{upper}{The upper boundary of the boostraped confidence interval.}
+#'  }
 #' @note
 #'  Ramanujan approximation is used for \eqn{x!} computation if \eqn{x > 170}.
 #' @references
@@ -481,36 +489,22 @@ setGeneric(
 #'  \code{\link{turnover}}
 #'  \code{\link{similarity}}
 #' @docType methods
-#' @name diversity
-#' @rdname diversity
+#' @name heterogeneity-index
+#' @rdname heterogeneity-index
 NULL
 
-#' @rdname diversity
-#' @aliases diversity-method
+#' @rdname heterogeneity-index
+#' @aliases index_heterogeneity-method
 setGeneric(
-  name = "diversity",
-  def = function(object, ...) standardGeneric("diversity")
+  name = "index_heterogeneity",
+  def = function(object, ...) standardGeneric("index_heterogeneity")
 )
 
-#' @rdname diversity
[…]
