+++ R/show.R
@@ -35,6 +35,16 @@ setMethod(
   }
 )
 
+# DiversityIndex ===============================================================
+setMethod(
+  f = "show",
+  signature = "DiversityIndex",
+  definition = function(object) {
+    cat(sprintf("<%s: %s>\n", class(object), object@method))
+    print(object@index)
+  }
+)
+
 # Logical matrix ===============================================================
 setMethod(
   f = "show",
