+++ codecov.yml
@@ -1,5 +1,8 @@
 comment: false
 
+ignore:
+  - "R/deprecate.R"
+
 coverage:
   status:
     project:
