### 2019-11-20 cb680d5
Add a paragraph concerning the state of the field (#15)
