+++ paper.bib
@@ -20,6 +20,16 @@
 	date = {1977},
 	author = {Bertin, Jacques}
 }
+@article{dray2007,
+  title = {The {ade4} Package: Implementing the Duality Diagram for Ecologists},
+  author = {Dray, St\'ephane and Dufour, Anne-B\'eatrice},
+  journal = {Journal of Statistical Software},
+  year = {2007},
+  volume = {22},
+  number = {4},
+  pages = {1-20},
+  doi = {10.18637/jss.v022.i04},
+}
 @article{dunnell1970,
 	langid = {english},
 	title = {Seriation {{Method}} and {{Its Evaluation}}},
@@ -90,6 +100,13 @@
   publisher = {{Springer}},
   date = {2002},
   author = {O'Brien, Michael J and Lyman, R. Lee}
+}
+@manual{oksanen2019,
+  title = {vegan: Community Ecology Package},
+  author = {Jari Oksanen and F. Guillaume Blanchet and Michael Friendly and Roeland Kindt and Pierre Legendre and Dan McGlinn and Peter R. Minchin and R. B. O'Hara and Gavin L. Simpson and Peter Solymos and M. Henry H. Stevens and Eduard Szoecs and Helene Wagner},
+  year = {2019},
+  note = {R package version 2.5-6},
+  url = {https://CRAN.R-project.org/package=vegan},
 }
 @article{petrie1899,
   title = {Sequences in Prehistoric Remains},
