+++ paper.md
@@ -28,13 +28,14 @@ The quantitative analysis of archaeological assemblages can thus be carried out 
 
 # Summary
 
-`tabula` provides a convenient and reproducible toolkit for analyzing, seriating, and visualizing archaeological count data, e.g., artifacts, faunal remains, etc.
+[`tabula`](https://CRAN.R-project.org/package=tabula) provides a convenient and reproducible toolkit for analyzing, seriating, and visualizing archaeological count data, e.g., artifacts, faunal remains, etc. 
+
+Several R packages, e.g. [`ade4`](https://CRAN.R-project.org/package=ade4) [@dray2007] or [`vegan`](https://CRAN.R-project.org/package=vegan) [@oksanen2019], allow the estimation of diversity indices and implements seriation/ordination methods, but these packages are mainly oriented towards ecological issues. `tabula` provides archaeologically-orientated implementations that allow the integration of specific data (dates, stratigraphy, etc.) and
+offers a consistent framework. The latter is of particular value since `tabula` is designed to be used both by archaeologists and by students with little background in courses on dating methods and applied statistics in archaeology.
 
 The package uses a set of S4 classes for archaeological data matrices that extend the `matrix` data type. These new classes represent different specialized matrices: incidence, abundance, co-occurrence, and (dis)similarity. Methods for a variety of functions applied to objects from these classes provide tools for relative and absolute dating and analysis of (chronological) patterns.
 
-`tabula` includes functions for matrix seriation (`seriate_*`), as well as chronological modeling and dating (`date_*`) of archaeological assemblages and objects. Resulting models can be checked for stability and refined with resampling methods (`refine_*`). Estimated dates can then be displayed as tempo or activity plot [@dye2016] to assess rhythms over long periods. Beyond these, `tabula` provides several tests (`test_*`) and measures of diversity within and between archaeological assemblages: heterogeneity and evenness (Brillouin, Shannon, Simpson, etc.), richness and rarefaction (Chao1, Chao2, ACE, ICE, etc.), turnover and similarity (Brainerd-Robinson, etc.). Finally, the package makes it easy to visualize count data and statistical thresholds (`plot_*`): rank vs. abundance plots, heatmaps, @ford1962, and @bertin1977 diagrams.
-
-`tabula` is designed to be used both by archaeologists and by students in courses on dating methods and applied statistics in archaeology.
+`tabula` includes functions for matrix seriation (`seriate_*`), as well as chronological modeling and dating (`date_*`) of archaeological assemblages and objects. Resulting models can be checked for stability and refined with resampling methods (`refine_*`). Estimated dates can then be displayed as tempo or activity plot [@dye2016] to assess rhythms over long periods. Beyond these, `tabula` provides several tests (`test_*`) and measures of diversity within and between archaeological assemblages (`index_*`): heterogeneity and evenness (Brillouin, Shannon, Simpson, etc.), richness and rarefaction (Chao1, Chao2, ACE, ICE, etc.), turnover and similarity (Brainerd-Robinson, etc.). Finally, the package makes it easy to visualize count data and statistical thresholds (`plot_*`): rank vs. abundance plots, heatmaps, @ford1962, and @bertin1977 diagrams.
 
 # Acknowledgements
 All the contributors have made it possible to develop this project through their helpful discussion and by bringing in new ideas: Jean-Baptiste Fourvel, Brice Lebrun, Ben Marwick, Matthew Peeples, and Anne Philippe.
