+++ NAMESPACE
@@ -26,6 +26,7 @@ exportMethods(get_features)
 exportMethods(get_id)
 exportMethods(get_order)
 exportMethods(get_totals)
+exportMethods(index_composition)
 exportMethods(index_evenness)
 exportMethods(index_heterogeneity)
 exportMethods(index_richness)

[…]
