+++ paper.md
@@ -30,7 +30,7 @@ The quantitative analysis of archaeological assemblages can thus be carried out 
 
 [`tabula`](https://CRAN.R-project.org/package=tabula) provides a convenient and reproducible toolkit for analyzing, seriating, and visualizing archaeological count data, e.g., artifacts, faunal remains, etc. 
 
-Several R packages, e.g. [`ade4`](https://CRAN.R-project.org/package=ade4) [@dray2007] or [`vegan`](https://CRAN.R-project.org/package=vegan) [@oksanen2019], allow the estimation of diversity indices and implements seriation/ordination methods, but these packages are mainly oriented towards ecological issues. `tabula` provides archaeologically-orientated implementations that allow the integration of specific data (dates, stratigraphy, etc.) and
+Several R packages, e.g. [`ade4`](https://CRAN.R-project.org/package=ade4) [@dray2007], [`SpadeR`](https://CRAN.R-project.org/package=SpadeR) [@chao2016] or [`vegan`](https://CRAN.R-project.org/package=vegan) [@oksanen2019], allow the estimation of diversity indices and implements seriation/ordination methods, but these packages are mainly oriented towards ecological issues. `tabula` provides archaeologically-orientated implementations that allow the integration of specific data (dates, stratigraphy, etc.) and
 offers a consistent framework. The latter is of particular value since `tabula` is designed to be used both by archaeologists and by students with little background in courses on dating methods and applied statistics in archaeology.
 
 The package uses a set of S4 classes for archaeological data matrices that extend the `matrix` data type. These new classes represent different specialized matrices: incidence, abundance, co-occurrence, and (dis)similarity. Methods for a variety of functions applied to objects from these classes provide tools for relative and absolute dating and analysis of (chronological) patterns.
