+++ paper.bib
@@ -20,6 +20,13 @@
 	date = {1977},
 	author = {Bertin, Jacques}
 }
+@manual{chao2016,
+  title = {SpadeR: Species-Richness Prediction and Diversity Estimation with R},
+  author = {Anne Chao and K. H. Ma and T. C. Hsieh and Chun-Huo Chiu},
+  year = {2016},
+  note = {R package version 0.1.1},
+  url = {https://CRAN.R-project.org/package=SpadeR},
+}
 @article{dray2007,
   title = {The {ade4} Package: Implementing the Duality Diagram for Ecologists},
   author = {Dray, St\'ephane and Dufour, Anne-B\'eatrice},
