### 2019-11-25 8765578
Fix missing ggplot2 namespace
