+++ NAMESPACE
@@ -57,11 +57,25 @@ exportMethods(test_fit)
 exportMethods(turnover)
 importFrom(ca,ca)
 importFrom(ggplot2,aes)
+importFrom(ggplot2,coord_fixed)
+importFrom(ggplot2,element_blank)
+importFrom(ggplot2,element_rect)
+importFrom(ggplot2,element_text)
+importFrom(ggplot2,facet_wrap)
+importFrom(ggplot2,geom_line)
 importFrom(ggplot2,geom_path)
 importFrom(ggplot2,geom_point)
+importFrom(ggplot2,geom_tile)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,labs)
+importFrom(ggplot2,scale_size_area)
+importFrom(ggplot2,scale_x_continuous)
+importFrom(ggplot2,scale_x_discrete)
 importFrom(ggplot2,scale_x_log10)
+importFrom(ggplot2,scale_y_continuous)
+importFrom(ggplot2,scale_y_discrete)
+importFrom(ggplot2,scale_y_log10)
+importFrom(ggplot2,theme)
 importFrom(methods,as)
 importFrom(methods,callNextMethod)
 importFrom(methods,new)
