### 2019-11-25 f4c1b16
Fix typo (openjournals/joss-reviews#1821)
