+++ R/tabula-package.R
@@ -2,16 +2,17 @@
 #' \tabular{ll}{
 #'  \strong{Package:} \tab tabula \cr
 #'  \strong{Type:} \tab Package \cr
-#'  \strong{Version:} \tab 1.3.0 \cr
+#'  \strong{Version:} \tab 1.4.0 \cr
 #'  \strong{License:} \tab GPL-3 \cr
-#'  \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.1489944}{10.5281/zenodo.1489944}
+#'  \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.1489944}{10.5281/zenodo.1489944} \cr
+#'  \strong{JOSS:} \tab \href{https://doi.org/10.21105/joss.01821}{10.21105/joss.01821} \cr
 #' }
 #' @author
 #' \strong{Full list of authors and contributors} (alphabetic order)
 #'
 #' \tabular{ll}{
-#'  Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
 #'  Jean-Baptiste Fourvel \tab \emph{CNRS, France} \cr
+#'  Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
 #'  Brice Lebrun \tab \emph{Aix-Marseille Université, France} \cr
 #'  Ben Marwick \tab \emph{University of Washington, USA} \cr
 #'  Matthew Peeples \tab \emph{Arizona State University, USA} \cr
@@ -38,9 +39,12 @@
 # See https://github.com/tidyverse/magrittr/issues/29
 if(getRversion() >= "2.15.1") utils::globalVariables(c("."))
 
+#' @importFrom ca ca
+#' @importFrom ggplot2 ggplot aes coord_fixed element_blank element_rect
+#' element_text facet_wrap geom_line geom_path geom_point geom_tile labs
+#' scale_size_area scale_x_continuous scale_x_discrete scale_x_log10
+#' scale_y_continuous scale_y_discrete scale_y_log10 theme
 #' @importFrom methods as callNextMethod new setClass setClassUnion setGeneric
 #' setMethod validObject
-#' @importFrom ca ca
-#' @importFrom ggplot2 ggplot aes geom_path geom_point labs scale_x_log10
 #' @importFrom rlang .data
 NULL
