+++ NEWS.md
@@ -1,163 +1,128 @@
+## tabula 1.4.0
 
-
-
-
-
-<!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
-
-<!-- ## tabula 1.3.0 (2019-09-20) -->
-
-## tabula 1.3.0
+* Published in the [*Journal of Open Source Software*](https://doi.org/10.21105/joss.01821).
 
 ### New classes and methods
-
-  - `Matrix` this S4 class is now the superclass of all matrix-like
-    classes.
-  - `AbundanceMatrix` this virtual S4 class is defined as the superclass
-    of `CountMatrix`, `FrequencyMatrix` and `IncidenceMatrix`.
-  - `SpaceTime` this S4 class represents space-time informations.
-  - `as_*()` coerce a `matrix` or `data.frame` to a `CountMatrix`,
-    `FrequencyMatrix`, `IncidenceMatrix`, `OccurrenceMatrix` or
-    `SimilarityMatrix`.
-  - `date_event()` replaces `dateEvent()`.
-  - `date_mcd()` allows Mean Ceramic Date estimation.
-  - `get_dates()` and `set_dates<-` allow to extract and replace
-    chronological informations in `AbundanceMatrix` objects.
-  - `plot_bertin()` and `plot_ford()` replace `plotBar()`.
-  - `plot_date()` replaces `plotDate()`.
-  - `plot_date()` gained a method for `AbundanceMatrix` objects.
-  - `plot_heatmap()` replaces `plotMatrix()`.
-  - `plot_rank()` replaces `plotRank()`.
-  - `plot_spot()` replaces `plotSpot()`.
-  - `plot_time()` produces an abundance *vs.* time graph.
-  - `refine_dates()` and `refine_seriation()` replace `refine()`.
-  - `seriate_reciprocal()` and `seriate_correspondance()` replace
-    `seriate()`.
-  - `test_diversity()` allows Shannon diversity test.
-  - `test_fit()` produces a Frequency Increment Test.
+* `DiversityIndex`, `HeterogeneityIndex`, `EvennessIndex` and `RichnessIndex` S4 classes represent diversity index.
+* `index_heterogeneity()` replaces `diversity()`.
+* `index_evenness()` replaces `evenness()`.
+* `index_richness()` replaces `richness()`.
+* `index_composition()` allows to estimate asymptotic species richness.
+* `plot_diversity()` produces a diversity *vs.* sample size graph and allow to compare estimates with simulated assemblages.
+* Add replacement methods for the `*Matrix` classes.
 
 ### Bugfixes & changes
-
-  - `CountMatrix`, `FrequencyMatrix` and `IncidenceMatrix` now also
-    contain the `SpaceTime` class.
-  - Deprecate `plotBar()`, `plotMatrix()`, `plotRank()`, `plotSpot()`,
-    `refine()`, `seriate()`.
-  - Remove `dateEvent()`.
-  - Empty rows/columns are removed prior to CA seriation to avoid error
-    in `svd()`.
-
-### Enhancements
-
-  - Add the Merzbach ceramics dataset.
-  - The `plot_date()` method for `DateModel` objects now allows to
-    display an activity or a tempo plot.
+* Deprecate `diversity()`, `evenness()` and `richness()`.
 
 ### Internals
+* Display progress bars only if `interactive()` is `TRUE` and {pbapply} is installed.
 
-  - Reduce required R version to 3.2.
-  - Error handling has been revised and error messages have been
-    harmonized.
-  - Refer to {ggplot2} functions using `::` (stop importing the entire
-    package).
-  - Use {vdiffr} to test graphical output.
-  - Replace `FactoMinerR::CA()` with `ca::ca()` (this avoids having to
[…]
