### 2019-11-25 #1821 @nfrerebeau
I don't know what number you need:

* DOI of version 1.4.0 (including changes made during the review): 10.5281/zenodo.3552904
* DOI for all versions: 10.5281/zenodo.1489944
