### 2019-11-25 #1821 @oliviaguest
@nfrerebeau can you check the final proof and add any missing DOIs that apply, please? 😊
