### 2019-11-25 e4627c0
Add missing DOI (openjournals/joss-reviews#1821)
