+++ paper.bib
@@ -116,9 +116,12 @@
   url = {https://CRAN.R-project.org/package=vegan},
 }
 @article{petrie1899,
-  title = {Sequences in Prehistoric Remains},
+  langid = {english},
+  title = {Sequences in {{Prehistoric Remains}}},
   volume = {29},
-  journaltitle = {Journal of the Anthropological Institute},
+  doi = {10.2307/2843012},
+  number = {3/4},
+  journaltitle = {The Journal of the Anthropological Institute of Great Britain and Ireland},
   date = {1899},
   pages = {295-301},
   author = {Petrie, W. M. F.}
