+++ vignettes/bibliography.bib
@@ -1,779 +1,868 @@
 
 @article{bellanger2006,
-	title = {Statistical Aspects of Pottery Quantification for the Dating of Some Archaeological Contexts},
-	volume = {48},
-	issn = {0003813X},
-	doi = {10.1111/j.1475-4754.2006.00249.x},
-	pages = {169--183},
-	number = {1},
-	journaltitle = {Archaeometry},
-	author = {Bellanger, L. and Husi, Ph. and Tomassone, R.},
-	date = {2006-02}
+  langid = {english},
+  title = {Statistical {{Aspects}} of {{Pottery Quantification}} for the {{Dating}} of {{Some Archaeological Contexts}}},
+  volume = {48},
+  issn = {0003813X},
+  doi = {10.1111/j.1475-4754.2006.00249.x},
+  number = {1},
+  journaltitle = {Archaeometry},
+  date = {2006-02},
+  pages = {169-183},
+  author = {Bellanger, L. and Husi, Ph. and Tomassone, R.}
 }
 
 @article{bellanger2006a,
-	title = {Une approche statistique pour la datation de contextes archéologiques},
-	volume = {54},
-	issn = {0003813X},
-	doi = {10.1111/j.1475-4754.2006.00249.x},
-	pages = {65--81},
-	number = {2},
-	journaltitle = {Revue de statistique appliquée},
-	author = {Bellanger, L. and Husi, Ph. and Tomassone, R.},
-	date = {2006}
+  langid = {french},
+  title = {Une approche statistique pour la datation de contextes arch\'eologiques},
+  volume = {54},
+  issn = {0003813X},
+  doi = {10.1111/j.1475-4754.2006.00249.x},
+  number = {2},
+  journaltitle = {Revue de statistique appliqu\'ee},
+  date = {2006},
+  pages = {65-81},
+  author = {Bellanger, L. and Husi, Ph. and Tomassone, R.}
 }
 
 @article{bellanger2008,
-	title = {A Statistical Approach for Dating Archaeological Contexts},
-	volume = {6},
-	pages = {135--154},
-	journaltitle = {Journal of Data Science},
-	author = {Bellanger, L. and Tomassone, R. and Husi, P.},
-	date = {2008}
+  langid = {english},
+  title = {A {{Statistical Approach}} for {{Dating Archaeological Contexts}}},
+  volume = {6},
+  journaltitle = {Journal of Data Science},
+  date = {2008},
+  pages = {135-154},
+  author = {Bellanger, L. and Tomassone, R. and Husi, P.}
 }
 
 @article{bellanger2012,
-	title = {Statistical Tool for Dating and Interpreting Archaeological Contexts Using Pottery},
-	volume = {39},
-	issn = {03054403},
-	doi = {10.1016/j.jas.2011.06.031},
-	pages = {777--790},
-	number = {4},
-	journaltitle = {Journal of Archaeological Science},
-	author = {Bellanger, Lise and Husi, Philippe},
-	date = {2012-04}
+  langid = {english},
+  title = {Statistical {{Tool}} for {{Dating}} and {{Interpreting Archaeological Contexts Using Pottery}}},
+  volume = {39},
+  issn = {03054403},
+  doi = {10.1016/j.jas.2011.06.031},
+  number = {4},
+  journaltitle = {Journal of Archaeological Science},
+  date = {2012-04},
+  pages = {777-790},
+  author = {Bellanger, Lise and Husi, Philippe}
 }
 
 @inproceedings{bellanger2013,
-	location = {Paris},
-	title = {Mesurer et modéliser le temps inscrit dans la matière à partir d'une source matérielle : la céramique médiévale},
[…]
