### 2019-11-25 #1821 @nfrerebeau
@oliviaguest only one of the two references had a DOI missing. I just added it!
