### 2019-11-25 #1821 @nfrerebeau
@oliviaguest yes, the suggested DOI points to a review of the article and not the original publication. As far as I know, this reference does not have a DOI.
