### 2019-12-01 #1821 @oliviaguest
@labarba :wave: I assumed tagging @openjournals/joss-eics above is enough to get this accepted... What am I missing?
