### 2019-12-01 #1821 @Kevin-Mattheus-Moerman
@nfrerebeau below are some suggested changes for your paper:

- [ ] The paper contains 1 instance of `...allowed the...` and 3 instances of `...allow the...`. Please consider adding `for`, i.e. so it reads `... allowed for the...`, and `... allow for the...` (or change to use `enabled`, `enables`).
- [ ] In this sentence the referencing can be improved `It may use a priori information, e.g., absolute dates or stratigraphical constraints: Poblome & Groenen (2003), and allows the analysis of chronological patterns in a socio-economic or cultural perspective, e.g., Bellanger & Husi (2012), Lipo, Madsen, & Dunnell (2015).`,
Consider reformatting to:
`It may use a priori information, e.g., absolute dates or stratigraphical constraints (Poblome & Groenen 2003), and allows for the analysis of chronological patterns in a socio-economic or cultural perspective (e.g., (Bellanger & Husi 2012), (Lipo, Madsen, & Dunnell 2015)).`
- [ ] In `...count data, e.g., artifacts, faunal remains, etc.` consider rephrasing to `...count data, such as artifacts and faunal remains`, which reads better and avoids the use of `etc.` at the end of a sentence.
- [ ] In `...of diversity indices and implements seriation/ordination methods, but...`, since packages is plural this should have `implement` not `implements`.
- [ ] Should the `[a]` be added here? `...Estimated dates can then be displayed as [a] tempo or activity plot...`
- [ ] If I understand the following correctly:  `...abundance plots, heatmaps, Ford (1962), and Bertin (1977) diagrams.` , perhaps it should read this instead: `...abundance plots, heatmaps, and Ford (1962) and Bertin (1977) diagrams.`
- [ ] In `All the contributors have made...`, I suggest: `The following contributors have made...`
