### 2019-12-01 #1821 @Kevin-Mattheus-Moerman
@nfrerebeau
- [ ] Please amend the meta data for the Zenodo archived version such that the title and the author set matches that of the paper.
- [ ] The archived version lists version v1.4.0 (which I've updated here already), can you confirm this is correct (and that this is the version of the reviewed and archived software)?
