### 2019-12-02 6b764c0
Rephrase and reformat parts of the paper (openjournals/joss-reviews #1821)
