+++ paper.md
@@ -22,22 +22,22 @@ bibliography: paper.bib
 
 Detecting and quantifying material and cultural variations in time and space are important methodological issues in archaeology. To solve these issues, we need to construct reliable chronologies and quantitative descriptions of archaeological assemblages, i. e. archaeological sites or intrasite units, each described as a set of $p$ different objects.
 
-Building chronologies involves distinguishing between relative (providing only a chronological sequence) and absolute dating methods (that yield calendric indicators) [@obrien2002]. Within relative dating, matrix seriation is a long-established method---it was first formulated by @petrie1899---and has allowed the construction of reference chronologies [@ihm2005]. For a set $X$ of $n$ archaeological assemblages, the seriation problem comes down to discovering in $X$ an order inferred as chronological. This approach relies on a set of well-defined statistical and archaeological assumptions [@dunnell1970]. It may use *a priori* information, e.g., absolute dates or stratigraphical constraints: @poblome2003, and allows the analysis of chronological patterns in a socio-economic or cultural perspective, e.g., @bellanger2012, @lipo2015.
+Building chronologies involves distinguishing between relative (providing only a chronological sequence) and absolute dating methods (that yield calendric indicators) [@obrien2002]. Within relative dating, matrix seriation is a long-established method---it was first formulated by @petrie1899---and has allowed for the construction of reference chronologies [@ihm2005]. For a set $X$ of $n$ archaeological assemblages, the seriation problem comes down to discovering in $X$ an order inferred as chronological. This approach relies on a set of well-defined statistical and archaeological assumptions [@dunnell1970]. It may use *a priori* information, e.g., absolute dates or stratigraphical constraints [@poblome2003] and allows for the analysis of chronological patterns in a socio-economic or cultural perspective (e.g., [@bellanger2012], [@lipo2015]).
 
 The quantitative analysis of archaeological assemblages can thus be carried out in a synchronic, e.g., diversity measurements, or diachronic, e.g., evolutionary studies: selection process, patterns of cultural transmission, etc., way. These approaches cover a wide range of applications and have led to the development of a multitude of statistical models, but none have been systematically implemented to enable the deployment of reproducible workflows.
 
 # Summary
 
-[`tabula`](https://CRAN.R-project.org/package=tabula) provides a convenient and reproducible toolkit for analyzing, seriating, and visualizing archaeological count data, e.g., artifacts, faunal remains, etc. 
+[`tabula`](https://CRAN.R-project.org/package=tabula) provides a convenient and reproducible toolkit for analyzing, seriating, and visualizing archaeological count data, such as artifacts and faunal remains. 
 
-Several R packages, e.g. [`ade4`](https://CRAN.R-project.org/package=ade4) [@dray2007], [`SpadeR`](https://CRAN.R-project.org/package=SpadeR) [@chao2016] or [`vegan`](https://CRAN.R-project.org/package=vegan) [@oksanen2019], allow the estimation of diversity indices and implements seriation/ordination methods, but these packages are mainly oriented towards ecological issues. `tabula` provides archaeologically-orientated implementations that allow the integration of specific data (dates, stratigraphy, etc.) and
[…]
