### 2019-12-02 #1821 @nfrerebeau
@Kevin-Mattheus-Moerman I made the changes you suggested. Thank you for taking the time to improve my text.

I modified the title of the Zenodo archive. The author set in the meta data already matches that of the paper, or am I missing something?

I confirm that v1.4.0 is the version of the reviewed and archived software.
