### 2019-12-02 #1821 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
